requirements.txt
google-genai
chromadb
feedparser
duckduckgo-search
gradio
python-dotenv
aiohttp
newsapi-python
requests